Use a more readable Cython implementation.
Note that I kept the `@cython.nogil` modifiers, even though the feature is not actually used. It might still be relevant for readers that this implementation _is_ actually usable in multi-threaded code.

I'm also making use of the fact that, after copying the array, it's definitely contiguous (`double[::1]`). That brings another 5% speedup on my side.

Also, I'd rather give `l` and `r` explicit names (`left_child` and `right_child`) than using functions to calculate them. But that's not a performance question (the functions get inlined either way). Rather, it's a matter of better readability.

You may want to update the performance numbers.